requirements.txt
streamlit==1.32.2
extra-streamlit-components==0.1.71
sqlalchemy==2.0.29
openai==1.14.3
PyYAML~=6.0.1
langchain==0.1.13
decouple==0.0.7
langchain-community==0.0.29